Refactoring, modernization the entire project to match ES6 module style
