--- examples/config.js
@@ -1,12 +1,0 @@
-let config = {};
-
-config.mixin = {
-    client_id         : "********-8095-4960-8dd1-edf6e583a2a9",
-    aeskey            : "********7P3k172oBLW3g/TUZa6Xa5MrgOOzKfXdv5A=",
-    pin               : "****23",
-    session_id    : "********-eb33-4112-b30a-2ae287dfbe32",
-    privatekey    : "mixin_dev.key"
-}
-
-
-module.exports = config;
